% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_parser.R
\name{add_datatransform}
\alias{add_datatransform}
\title{Internal function to build and add the data transform function}
\usage{
add_datatransform(
  tpl,
  dataTransform = tpl$dataTransform,
  dataTransformParams = tpl$dataTransformParams,
  transform_options = NULL
)
}
\arguments{
\item{tpl}{a graphTemplate}

\item{dataTransform}{In string form, the name of the data transform function}

\item{dataTransformParams}{Parameters to pass to the dataTransform function}

\item{transform_options}{optional options for the data transform function}
}
\value{
a graphtemplate with data transform function
}
\description{
Internal function to build and add the data transform function
}
