% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTemplate_class.R
\name{addTernaryOrnaments}
\alias{addTernaryOrnaments}
\alias{addTernaryOrnaments.ternary}
\alias{rotateTernaryTemplate}
\alias{rotateTernaryTemplate.ternary}
\title{addTernary}
\usage{
addTernaryOrnaments(self, scale, rotation)

\method{addTernaryOrnaments}{ternary}(
  self,
  axes = T,
  apicesNames = T,
  ticks = F,
  grid = F,
  padding = c(0.03, 0.03, 0.03, 0.05)
)

rotateTernaryTemplate(self, rotation, scale, padding, setup = F)

\method{rotateTernaryTemplate}{ternary}(self, rotation = 0, scale = 1, padding, setup = F)
}
\arguments{
\item{self}{a template}

\item{scale}{scaling factor}

\item{rotation}{Angle to rotate}
}
\description{
Add ternary information (axes) to a template

Add ternary information (axes) to a template
}
