% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTemplate_class.R
\name{addTernaryOrnaments.graphTemplate}
\alias{addTernaryOrnaments.graphTemplate}
\title{Add axes to a ternary template}
\usage{
\method{addTernaryOrnaments}{graphTemplate}(self, scale, rotation)
}
\arguments{
\item{self}{a graphTemplate}

\item{scale}{The scale of the triangle: normally from 0 to 1 but in some contexts
the traingle could be defined as 0-100 (\%). Probably deprecated.}

\item{rotation}{in degrees, angle to rotate the graph (in progress).}
}
\description{
Add axes to a ternary template
}
\details{
This method is used to add ternary (pseudo) axes to a graphTemplate. The plotter
should remove the true axes (and ensure a square aspect ratio). The function
eventually will be able to rotate a ternary graph by an arbitrary amount.
}
