% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTemplate_class.R
\name{addTernaryOrnaments}
\alias{addTernaryOrnaments}
\alias{addTernaryOrnaments.graphTemplate}
\alias{addTernaryOrnaments.ternary}
\title{Add axes to a ternary template}
\usage{
addTernaryOrnaments(self, ...)

\method{addTernaryOrnaments}{graphTemplate}(self, ...)

\method{addTernaryOrnaments}{ternary}(
  self,
  axes = T,
  lwd.ax = 1,
  lty.ax = "solid",
  col.ax = "black",
  apicesNames = T,
  ticks = F,
  length.ti = 0.03,
  interval.ti = 10,
  lwd.ti = 1,
  lty.ti = "solid",
  col.ti = "black",
  grid = F,
  interval.gr = 10,
  lwd.gr = 0.5,
  lty.gr = "dotted",
  col.gr = "grey",
  padding = c(0.03, 0.03, 0.03, 0.05)
)
}
\arguments{
\item{self}{a graphTemplate}

\item{...}{to ternary method}

\item{axes, ticks, grid, apicesNames}{: should we add the relevant elements?}

\item{lwd.ax, lty.ax, col.ax}{graphical properties of the pseudoaxes}

\item{length.ti, interval.ti, lwd.ti, lty.ti, col.ti}{spacing between ticks, length, and their graphical properties}

\item{interval.gr, lwd.gr, lty.gr, col.gr}{spacing between grid lines, and their graphical properties}

\item{padding}{Extra space to add around the plot, to allow depicting the apices labels (left, top, right, bottom)}
}
\value{
updated version of self
}
\description{
Add axes to a ternary template
}
\details{
This method is used to add various elements to a ternary template, in particular
ternary (pseudo) axes, names of the apices, ticks and grid. The plotting function
should remove the true axes (and ensure a square aspect ratio).
Elements will honour the template rotation.
}
