% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{demoTrans}
\alias{demoTrans}
\title{Showing how to use options}
\usage{
demoTrans(wrdata, doubleBB = F)
}
\arguments{
\item{wrdata}{a WR matrix}

\item{doubleBB}{a tag that will do something...}
}
\value{
a matrix with columns AA and BB
@export
}
\description{
A demo function demonstrating how calculation options can be used.
}
\details{
this function serves no real purpose other than demonstrate how a
a data transformation function can take arguments. They can either be called
by passing them to the json parser, in transform_options; or defined in
the json itself in field dataTransformParams.
As always, user options will override the template.
}
