% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCDkit_additions.R
\name{FrostASI}
\alias{FrostASI}
\title{Calculate ASI based on Frost's definition}
\usage{
FrostASI(wrdata, originalDef = F, zeroP = T)
}
\arguments{
\item{wrdata}{A matrix with whole-rock composition, probably coming from GCDkit}

\item{originalDef}{Use Frost's original (incorrect) equation? (see details)}

\item{zeroP}{Boolean. If true, replace missing P2O5 values by na.}
}
\value{
a matrix with column ASI.
}
\description{
Calculate ASI based on Frost's definition
}
\details{
There is a well-known typo in Frost's paper, where the equation for
ASI is given with an incorrect coefficient (not respecting toechiometry).
This is generally acknowledged to be incorrect, but it is possible to
reinstall the original behaviour if wished.
}
