% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{FrostFeHook}
\alias{FrostFeHook}
\title{Hook function to change the position of the line in Frost Figure 1}
\usage{
FrostFeHook(self, wrdata, lbl, FeOonly = F)
}
\arguments{
\item{self}{The template}

\item{wrdata}{WR data}

\item{lbl}{labels}

\item{FeOonly}{if true, use FeO strictly (see details)}
}
\value{
modified template, wrdata, lbl
}
\description{
The position of the line is adjusted based on the Fe component
}
\details{
Frost's Fe nbr is said to use FeO. It is however unclear whether this
should be FeO strictly, or FeOt (all iron as FeO).
}
