% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCDkit_additions.R
\name{FrostFeNbr}
\alias{FrostFeNbr}
\title{Calculate Fe nbr for Frost plot}
\usage{
FrostFeNbr(wrdata, FeOonly = F)
}
\arguments{
\item{wrdata}{A matrix with whole-rock composition, probably coming from GCDkit}

\item{FeOonly}{if true, use FeO strictly (see details)}
}
\value{
a matrix with column FeNbr
}
\description{
Calculate Fe nbr for Frost plot
}
\details{
Frost's Fe nbr is said to use FeO. It is however unclear whether this
should be FeO strictly, or FeOt (all iron as FeO).
NB in this case we should change the template depending on the option!!
}
