% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fct.R
\name{GCDkitToTibble}
\alias{GCDkitToTibble}
\title{Convert GCDkit-style data (WR and labels) into a tibble}
\usage{
GCDkitToTibble(gcdWR = WR, gcdlabels = get("labels", .GlobalEnv))
}
\arguments{
\item{gcdWR}{WR matrix from GCDkit}

\item{gcdlabels}{labels data.frame from GCDkit}
}
\value{
a tibble with all the data
}
\description{
Convert GCDkit-style data (WR and labels) into a tibble
}
\details{
this is a tidyverse function. It will fail if tidyverse is not loaded.
There is no explicit import in the function (to avoid creating a dependency during installation),
so errors may be cryptic, and probably related to magrittR pipe ("\%>\%")
}
