% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateElement_ggplot_methods.R
\name{gglayerTemplateElement}
\alias{gglayerTemplateElement}
\alias{gglayerTemplateElement.templateElement}
\alias{gglayerTemplateElement.lines}
\alias{gglayerTemplateElement.polygon}
\alias{gglayerTemplateElement.arrows}
\alias{gglayerTemplateElement.text}
\alias{gglayerTemplateElement.abline}
\alias{gglayerTemplateElement.points}
\alias{gglayerTemplateElement.curve}
\alias{gglayerTemplateElement.reservoirs}
\title{Turn a templateElement into a ggplot layer}
\usage{
gglayerTemplateElement(self)

\method{gglayerTemplateElement}{templateElement}(self)

\method{gglayerTemplateElement}{lines}(self)

\method{gglayerTemplateElement}{polygon}(self)

\method{gglayerTemplateElement}{arrows}(self)

\method{gglayerTemplateElement}{text}(self)

\method{gglayerTemplateElement}{abline}(self)

\method{gglayerTemplateElement}{points}(self)

\method{gglayerTemplateElement}{curve}(self)

\method{gglayerTemplateElement}{reservoirs}(self)
}
\arguments{
\item{self}{a template element}
}
\description{
This method converts a templateElement into a ggplot player, using
ggplot2::annotate (i.e. they are annotation layers)
It assumes that all defaults (col, lwd and lty at least)
are set properly.
}
