% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTemplate_class.R
\name{new_graphTemplate}
\alias{new_graphTemplate}
\alias{validate_graphTemplate}
\alias{graphTemplate}
\alias{print.graphTemplate}
\alias{print.unknown}
\alias{print.binary}
\alias{print.ternary}
\alias{print.plate}
\title{Class constructor for graphTemplate}
\usage{
new_graphTemplate(
  name = NULL,
  fullName = NULL,
  meta = NULL,
  Rdialect = NULL,
  dataTransform = NULL,
  hook = NULL,
  dataFilter = NULL,
  axesDefinition = NULL,
  axesName = NULL,
  log = log,
  limits = NULL,
  template = NULL
)

validate_graphTemplate(self)

graphTemplate(
  name = NULL,
  fullName = NULL,
  meta = NULL,
  Rdialect = NULL,
  dataTransform = function(z) {
     return(z)
 },
  hook = function(self, wrdata, lbl) {
     return(list(self = self, wrdata = wrdata, lbl
    = lbl))
 },
  dataFilter = NULL,
  axesDefinition = NULL,
  axesName = NULL,
  log = NULL,
  limits = NULL,
  template = NULL,
  diagramType = "unknown"
)

\method{print}{graphTemplate}(self, ...)

\method{print}{unknown}(self, ..)

\method{print}{binary}(self, ..)

\method{print}{ternary}(self, ..)

\method{print}{plate}(self, ..)
}
\arguments{
\item{All}{the elements of the list, as given above, correspond to parameters of the
class constructor.}
}
\description{
Class constructor for graphTemplate

Internal validation function for graphTemplate

A class that stores all the information allowing to reproduce a graph
}
\details{
A graphTemplate is a list, with the following elements:
\itemize{
\item a name (short, typically the name of the json file it is loaded from)
\item a fullName, more explicit
\item meta, itself a list containing various metadata such as a reference, template creator, etc.
\item Rdialect, when the diagram requires functions (for data transformation) belonging to
a certain package (gcdkit, probably). The plotter is expected to complain if the right library
is not loaded...
\item dataTransform a function that takes a data.frame and returns the plotting coordinates (it is
usually as simple as function(x){return(x)}, but more elaborate functions can be used. The
json parser is able to interpret, to a degree, what is found in its dataTransform field and combine it
with user-supplied arguments).
\item dataFilter. Some graphs plot only data obeying certain conditions, such as
SiO2 < 54 \%. This should be supplied as a string (fit for filtering by the plotter).
\item axesDefinition the axes, expressed in terms of the (col)names of the data
\item axesName as they will be displayed on the graph. Can be prettier... GCDkit will
attempt (using GCDkit::annotate) to prettify them. ggplot is not as clever (but it is
easy to modify an axis name).
\item log. "", "x", "y" or "xy" in usual R convention.
\item limits, a list with X and Y components, two vectors of two elements...
\item template the template itself, a collection of elements of type templateElement
}
In addition, a graphTemplate has a subclass, defined by diagramType: binary or ternary.
This will be interpreted by the plotter.
In fact, since a graphTemplate is a list, it can contain any other element. However,
the creator function takes only these arguments, and only them will be used for plotting...
Additionally, a graphTemplate can be of subclass binary or ternary.
}
