% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_parser.R
\name{load_json_file}
\alias{load_json_file}
\title{Reads a json template file into a list}
\usage{
load_json_file(json, path = NULL)
}
\arguments{
\item{json}{Name of the template file (if the .json extension is missing, it will be added automatically)}

\item{path}{Path to json file}
}
\description{
A graphTemplate can be defined as a json file. This function
load a json (from package's directory or user-supplied location),
performs some basic syntaxic checks and returns a list for further processing.
}
\details{
If no path is supplied, the function will look in the json_template
directory of the package folder, and its subdirectories. It will return
the first match, starting by the top-level folder and exploring
sub-folders alphabetically.
}
