% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_parser.R
\name{make_function}
\alias{make_function}
\title{Internal function to build a function}
\usage{
make_function(fnName, params = NULL, fullArgs = F)
}
\arguments{
\item{fnName}{In string form, the name of function}

\item{params}{Parameters to pass to the fnName function}

\item{fullArgs}{If true, the function generated will take self (a template),
wrdata and lbl as arguments. Otherwose it will take only wrdata.}
}
\value{
a function
}
\description{
Internal function to build a function
}
