% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_utilities.R
\name{make_line}
\alias{make_line}
\title{Internal function. Make a templateElement line between two points.}
\usage{
make_line(ptFrom, ptTo, col, lty, lwd)
}
\arguments{
\item{ptFrom, ptTo}{the start and end point}

\item{lwd, lty, col}{Usual style parameters}
}
\value{
a templateElement object corresponding to a line between the points
}
\description{
Internal function. Make a templateElement line between two points.
}
