% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_parser.R
\name{make_plate_template}
\alias{make_plate_template}
\title{Convert a raw list into a proper graphTemplate}
\usage{
make_plate_template(
  tpl_raw,
  path,
  plateMeta = c("details", "reference", "url", "templateAuthor", "templateConversion",
    "templateHistory"),
  template_options = NULL,
  transform_options = NULL,
  style_options = c(colDefault = "black", ltyDefault = "solid", lwdDefault = 1),
  doFilter = T,
  ...
)
}
\arguments{
\item{tpl_raw}{a list, that will be converted to a graphTemplate.
Probably loaded from json, using \link{load_json_file}}

\item{plateMeta}{the names of the fields from the json file that should be kept
as metadata for the plate}

\item{template_options}{Named vector. They contain the switches to activate, or desactivate,
controlling the display of some graph elements (see \link{jsonDiagramFormat}).}

\item{transform_options}{Additional parameters to pass to the data transformation function}

\item{style_options}{Options to control the styling of some elements of the template (see \link{jsonDiagramFormat}).}

\item{doFilter}{if the template contains a filter (eg SiO2 > 45), should we respect it?}

\item{...}{Further arguments passed to make_single_template(),
i.e. arguments applying to individual templates.}
}
\description{
Convert a raw list into a proper graphTemplate
}
