% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTemplate_ggplot_methods.R
\name{makeggTemplate}
\alias{makeggTemplate}
\alias{makeggTemplate.graphTemplate}
\alias{makeggTemplate.binary}
\alias{makeggTemplate.ternary}
\title{Build an empty graph template}
\usage{
makeggTemplate(self, ...)

\method{makeggTemplate}{graphTemplate}(self, ...)

\method{makeggTemplate}{binary}(self, ...)

\method{makeggTemplate}{ternary}(self, ...)
}
\arguments{
\item{self}{a graphTemplate}

\item{...}{probably not used}
}
\description{
Build an empty graph template
}
\details{
This function assembles the "static" elements of a template (lines, text)
as well as axis and scale definitions and convert them for GCDkit. It does
not map aesthetics nor calculate data points, the user must do it in this case
(see examples). Alternately, use \link{plotgg} for a more global, but less
flexible solution.
}
