% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTemplate_figaro_methods.R
\name{makeName}
\alias{makeName}
\alias{makeName.graphTemplate}
\title{makeName}
\usage{
makeName(self, W)

\method{makeName}{graphTemplate}(self, W)
}
\arguments{
\item{self}{a template}

\item{W}{the axis to format}
}
\description{
Nice looking name for an axis
}
