% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{my_test_hook}
\alias{my_test_hook}
\title{Just testing!}
\usage{
my_test_hook(self, wrdata, lbl)
}
\arguments{
\item{self}{The template}

\item{wrdata}{WR data}

\item{lbl}{labels}
}
\description{
Just testing!
}
