% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{OhtaAraiParams}
\alias{OhtaAraiParams}
\title{Recalculate Otha + Arai M, F and W parameters}
\usage{
OhtaAraiParams(where)
}
\arguments{
\item{where}{A matrix with whole-rock composition, probably coming from GCDkit}
}
\value{
a matrix with columns M, F and W.
}
\description{
Recalculate Otha + Arai M, F and W parameters
}
