% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_parser.R
\name{parseJsonTemplate}
\alias{parseJsonTemplate}
\title{Parse a template loaded from json, converting the options as required}
\usage{
parseJsonTemplate(
  json,
  path = NULL,
  plateMeta = c("details", "reference", "url", "templateAuthor", "templateConversion",
    "templateHistory"),
  ...
)
}
\arguments{
\item{json}{Name of the template file (if the .json extension is missing, it will be added automatically)}

\item{path}{Path to json file. By default the package files.}

\item{plateMeta}{the names of the fields from the json file that should be kept
as metadata for the plate (if we are looking at a plate). Single plot-level metadata
should be defined using meta, and are passed to make_single_template via ...}

\item{...}{further options to be processed, see \link{make_single_template}}
}
\value{
a graphTemplate object.
}
\description{
This function combines a "raw" list, loaded from json file, with user options.
It performs the required modifications, such as changing colours or
removing elements, and returns a properly formatted graphTemplate object.
}
\details{
This is a thin wrapper on make_single_template. It main job is to decide whether
we are looking at a single plot, or a plate, and send the template or templates
to make_single_template
}
