% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCDkit_connector.R
\name{plotDiagram_json}
\alias{plotDiagram_json}
\title{A replacement for GCDkit's plotDiagram, with the same API}
\usage{
plotDiagram_json(
  diagram,
  select.samples = TRUE,
  new = TRUE,
  main = NULL,
  width = 6.5,
  height = 6.5,
  pointsize = 10,
  bg = "transparent",
  autoscale = TRUE,
  interactive = FALSE,
  ...
)
}
\description{
In principle, it should take the same arguments and return the same Figaro
graph, so it aims to be a transparent replacement.
}
