% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTemplate_ggplot_methods.R
\name{plotgg}
\alias{plotgg}
\alias{plotgg.graphTemplate}
\alias{plotgg.binary}
\alias{plotgg.ternary}
\title{Plot a template in ggplot2 context}
\usage{
plotgg(self, wrdata, lbl, new, ...)

\method{plotgg}{graphTemplate}(self, wrdata, lbl, new, ...)

\method{plotgg}{binary}(self, wrdata = WR, lbl = get("labels", .GlobalEnv), new, ...)

\method{plotgg}{ternary}(self, wrdata = WR, lbl = get("labels", .GlobalEnv), new, ...)
}
\arguments{
\item{self}{a graphTemplate}

\item{wrdata, lbl}{The usual GCDkit WR and labels objects, default to what is
in the global environment.}

\item{new}{should control whether a new window is opened, but ignored for now.}

\item{...}{probably not used}
}
\description{
Plot a template in ggplot2 context
}
\details{
The function that does the real job... this function converts the template
into ggplot objects and add points on it, obeying GCDkit-style conventions
(Symbol, Colour etc). If you are not coming from a GCDkit context and/or
want more granularity, it is better to use \link{makeggTemplate} and do the rest
by hand.

Symbols, colours, and size are taken from variables Symbol, Colour and Size in
lbl (itself potentially coming from GCDkit, or from the global variable labels).
}
