% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_parser.R
\name{process_template_options}
\alias{process_template_options}
\title{Internal function, modifies a template based on user options}
\usage{
process_template_options(
  tpl_objects,
  switching_options = NULL,
  style_options = c(colDefault = "black", ltyDefault = "solid", lwdDefault = 1)
)
}
\arguments{
\item{tpl_objects}{A list containing elements that can be interpreted as templateElement}

\item{switching_options}{Named vector. They contain the switches to activate, or desactivate,
controlling the display of some graph elements (see \link{jsonDiagramFormat}).}

\item{style_options}{Options to control the styling of some elements of the template (see \link{jsonDiagramFormat}).}
}
\value{
a list of templateElement, modified and "filtered" as required
}
\description{
Internal function, modifies a template based on user options
}
\details{
each element of the input list will be converted into a templateElement object.
If it is switched "off", it will be removed. If appropriate, its properties
(col, lty and lwd) will be modified.
}
