% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{projBiotHook}
\alias{projBiotHook}
\title{Hook function to tweak Moyen et al. proj from biotite}
\usage{
projBiotHook(self, wrdata, lbl)
}
\arguments{
\item{self}{The template}

\item{wrdata}{WR data}

\item{lbl}{labels}
}
\value{
modified template, wrdata, lbl
}
\description{
Hook function to tweak Moyen et al. proj from biotite
}
