% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{recastAnhydrous}
\alias{recastAnhydrous}
\title{Recalculate WR composition to anhydrous}
\usage{
recastAnhydrous(where)
}
\arguments{
\item{where}{A matrix with whole-rock composition, probably coming from GCDkit}
}
\value{
A matrix with columns of the same name, sum normalized to 100\% anhydrous.
}
\description{
As done in gCDkit (First.r).
Ultimately there will probably be a GCDkit function by that name !
}
