% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_utilities.R
\name{rotateCoordinates}
\alias{rotateCoordinates}
\title{Calculate coordinates following a rotation}
\usage{
rotateCoordinates(X, Y, rotation, scale)
}
\arguments{
\item{X, Y}{Original coordinates}

\item{rotation}{Angle, in degree}

\item{scale}{Scale factor}
}
\value{
data.frame with new X and Y (Xr, Yr)
}
\description{
Calculate coordinates following a rotation
}
