% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateElement_class.R
\name{rotateTemplateElement}
\alias{rotateTemplateElement}
\alias{rotateTemplateElement.templateElement}
\alias{rotateTemplateElement.text}
\alias{rotateTemplateElement.arrows}
\alias{rotateTemplateElement.abline}
\title{Rotate individual template elements}
\usage{
rotateTemplateElement(self, rotation, scale)

\method{rotateTemplateElement}{templateElement}(self, rotation = 0, scale = 1)

\method{rotateTemplateElement}{text}(self, rotation = 0, scale = 1)

\method{rotateTemplateElement}{arrows}(self, rotation = 0, scale = 1)

\method{rotateTemplateElement}{abline}(self, rotation = 0, scale = 1)
}
\arguments{
\item{self}{a template element}

\item{rotation}{Angle, in degree}

\item{scale}{Scale factor}
}
\description{
This method can be used to rotate a template element (when rotating
a ternary plot probably)
}
\details{
methods are implemented for
\itemize{
\item{} anything that has X and Y fields
\item{text}  (srt is also modified)
\item{arrows}  (x0, x1, y0 and y1 are modified)
\item{abline}  (converted into lines)
}
}
