% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTemplate_class.R
\name{rotateTernaryTemplate.graphTemplate}
\alias{rotateTernaryTemplate.graphTemplate}
\title{Add axes to a ternary template}
\usage{
\method{rotateTernaryTemplate}{graphTemplate}(self, rotation, scale, padding, setup = F)
}
\arguments{
\item{self}{a graphTemplate}

\item{rotation}{in degrees, angle to rotate the graph.}

\item{scale}{The scale of the triangle: normally from 0 to 1 but in some contexts
the triangle could be defined as 0-100 (\%). Probably deprecated.}

\item{padding}{Extra space to add on the sides, c(left,top,right,bottom)}

\item{setup}{If T, use during setup. Just rotate. If F, use when rotating an
existing plot. This will update self$ternaryRotation.}
}
\description{
Add axes to a ternary template
}
\details{
This method is used to add ternary (pseudo) axes to a graphTemplate. The plotter
should remove the true axes (and ensure a square aspect ratio). The function
eventually will be able to rotate a ternary graph by an arbitrary amount.
}
