% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateElement_class.R
\name{new_templateElement}
\alias{new_templateElement}
\alias{validate_templateElement}
\alias{templateElement}
\title{Constructor for templateElement class}
\usage{
new_templateElement(te, cls)

validate_templateElement(te)

templateElement(te, cls = te$plotFun)
}
\arguments{
\item{te}{An input list to convert to a template element}

\item{cls}{the class of the template element. Will be guessed from plotFun, if the
slot is present.}
}
\description{
A (graphical) element of a template

Each \link{graphTemplate} is made of various information, the core of
which is the "template" section. This contains individual elements (lines, texts, etc.)
that are all defined as a templateElement object.
templateElements have subclasses such as lines, text... defining how they will be treated.

Figaro does have functionalities to interpret and plot diverse elements, so no
further work is required (expect passing the template to Figaro, see ?plotFigaro).
In ggplot however, there is an extra step of converting the templateElement to
ggplot graphical primitives (ggproto).
}
