% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_utilities.R
\name{ternaryBoundingBox}
\alias{ternaryBoundingBox}
\title{The bounding box of a ternary diagram}
\usage{
ternaryBoundingBox(rotation, scale, padding = c(0.03, 0.03, 0.03, 0.05))
}
\arguments{
\item{rotation}{Angle, in degree}

\item{scale}{Scale factor}

\item{padding}{Extra space to add (as pct); left, top, right, botton}
}
\value{
A list with two components, X and Y.
}
\description{
The bounding box of a ternary diagram
}
