% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_utilities.R
\name{ternaryCoordinates}
\alias{ternaryCoordinates}
\title{Calculate ternary coordinates based on three variables, A, B and C}
\usage{
ternaryCoordinates(A, B, C, rotation = 0, scale = 1)
}
\arguments{
\item{A, B, C}{numeric vectors the values}

\item{rotation}{rotation angle, degrees}

\item{scale}{size of the triangle}
}
\value{
a five-column data.frame : X, Y, and the A, B and C coordinates (as supplied).
}
\description{
Calculate ternary coordinates based on three variables, A, B and C
}
