% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_utilities.R
\name{ternaryGrid}
\alias{ternaryGrid}
\title{Generate templateElements defining a grid}
\usage{
ternaryGrid(interval = 10, lwd = 0.5, lty = "dotted", col = "grey")
}
\arguments{
\item{interval}{Number of intervals on each axis. (interval-1) lines
will be drawn...}

\item{lwd, lty, col}{Usual parameters for the grid}
}
\value{
a list of templateElement objects defining the grid
}
\description{
Generate templateElements defining a grid
}
