% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_utilities.R
\name{ternaryTicks}
\alias{ternaryTicks}
\title{Generate templateElements defining ticks}
\usage{
ternaryTicks(
  length = 0.05,
  interval = 10,
  lwd = 1,
  lty = "solid",
  col = "black"
)
}
\arguments{
\item{length}{How far they should project in the diagram}

\item{interval}{Number of intervals on each axis. (interval-1) ticks
will be drawn...}

\item{lwd, lty, col}{Usual parameters for the ticks}
}
\value{
a list of templateElement objects defining ticks
}
\description{
Generate templateElements defining ticks
}
