% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fct.R
\name{theme_gcdkit}
\alias{theme_gcdkit}
\title{A ggplot theme that resembles GCDkit's}
\usage{
theme_gcdkit()
}
\description{
A ggplot theme that resembles GCDkit's
}
\details{
Works only in ggplot, obviously. If ggplot is not loaded, it does nothing...
}
