% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fct.R
\name{TibbleToGCDkit}
\alias{TibbleToGCDkit}
\title{(attempts to) convert a tibble into two GCDkit tables.}
\usage{
TibbleToGCDkit(WRD, nameCol = names(WRD)[1])
}
\arguments{
\item{WRD}{a tibble with all the info}

\item{column}{containing the sample names (ID), the first one by default}
}
\value{
a list with two components, WR and labels
}
\description{
(attempts to) convert a tibble into two GCDkit tables.
}
\details{
This function does its best, putting numerical data in WR and the
rest in labels, as per GCDkit convention
}
