% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTemplate_class.R
\name{addTernaryAxes}
\alias{addTernaryAxes}
\alias{addTernaryAxes.graphTemplate}
\alias{addTernaryAxes.ternary}
\title{addTernary}
\usage{
addTernaryAxes(self, scale, rotation)

\method{addTernaryAxes}{graphTemplate}(self, scale, rotation)

\method{addTernaryAxes}{ternary}(self, scale = 1, rotation = 0)
}
\arguments{
\item{self}{a graphTemplate}

\item{scale}{The scale of the triangle: normally from 0 to 1 but in some contexts
the traingle could be defined as 0-100 (\%). Probably deprecated.}

\item{rotation}{in degrees, angle to rotate the graph (in progress).}
}
\description{
Add ternary information (axes) to a template
}
\details{
This method is used to add ternary (pseudo) axes to a graphTemplate. The plotter
should remove the true axes (and ensure a square aspect ratio). The function
eventually will be able to rotate a ternary graph by an arbitrary amount.
}
