% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTemplate_class.R
\name{addTernaryOrnaments.graphTemplate}
\alias{addTernaryOrnaments.graphTemplate}
\title{Add axes to a ternary template}
\usage{
\method{addTernaryOrnaments}{graphTemplate}(self, ...)
}
\arguments{
\item{self}{a graphTemplate}

\item{...}{to ternary method}

\item{axes, ticks, grid}{: should we add the relevant elements?}

\item{interval.gr, lwd.gr, lty.gr, col.gr}{spacing between grid lines, and their graphical properties}

\item{length.ti, interval.ti, lwd.ti, lty.ti, col.ti}{spacing between ticks, length, and their graphical properties}

\item{lwd.ax, lty.ax, col.ax}{graphical properties of the pseudoaxes}

\item{padding}{Extra space to add around the plot, to allow depicting the apices labels (left, top, right, bottom)}
}
\value{
updated version of self
}
\description{
Add axes to a ternary template
}
\details{
This method is used to add various elements to a ternary template, in particular
ternary (pseudo) axes. The plotting function
should remove the true axes (and ensure a square aspect ratio).
Elements will honour the template rotation.
}
